# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Messenger automated tasks API",
    "version": "1.0.0",
    "description": "API documentation for messenger automated tasks",
    "contact": {
      "name": " email of developer",
      "email": "yankostov1@gmail.com"
    }
  },
  "tags": [
    {
      "name": "Task",
      "description": "Operations related to tasks"
    }
  ],
  "paths": {
    "/api/v1/tasks": {
      "get": {
        "tags": [
          "Task"
        ],
        "summary": "Get all tasks",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "type": "number",
            "description": "Page number to get"
          },
          {
            "in": "query",
            "name": "limit",
            "type": "number",
            "description": "Number of tasks on the page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/taskResponce"
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Task"
        ],
        "summary": "Create a task",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "task object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Task"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/taskResponce"
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      }
    },
    "/api/v1/tasks/{id}": {
      "get": {
        "tags": [
          "Task"
        ],
        "summary": "Get a task by ID",
        "operationId": "getTaskById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "number",
            "required": true,
            "description": "Id of the task to get"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/taskResponce"
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Task"
        ],
        "summary": "Update a task",
        "operationId": "updateTask",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "number",
            "required": true,
            "description": "Id of the task to update"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Task object to be updated",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Task"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/taskResponce"
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Task"
        ],
        "summary": "Delete a task by ID",
        "operationId": "deleteTask",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "number",
            "required": true,
            "description": "Id of the task to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/deletedTaskResponce"
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
    "taskResponce": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "datetime": {
          "type": "string"
        },
        "recurring": {
          "type": "string"
        },
        "chat_url": {
          "type": "string"
        },
        "bot_id": {
          "type": "integer"
        }
      }
    },
    "deletedTaskResponce": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "Task": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "datetime": {
          "type": "string"
        },
        "recurring": {
          "type": "string"
        },
        "chat_url": {
          "type": "string"
        },
        "bot_id": {
          "type": "string"
        }
      },
      "required": [
        "name",
        "message",
        "datetime",
        "recurring",
        "chat_url"
      ]
    },
    "InvalidResponse": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  }
}